Rename GlobalStyle to Style
Global Style 디렉터리 안에는 Global Style 이외에도 다른 스타일 엘리먼트를 함유하고 있는 파일들이 많습니다. 그리하여 수정하였습니다.
