+++ src/Constants/Styles/Detail.ts
